# activity-service/api/openapi.yaml
openapi: 3.1.0
info:
  title: FocusNest Activity Service
  version: 0.1.0
servers:
  - url: https://activity.focusnest.com
    description: Production
  - url: http://localhost:8080
    description: Local development
paths:
...
  /v1/productivities:
    get:
      operationId: listProductivities
      summary: List productivity entries for the authenticated user
      tags:
        - productivity
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: month
          description: Month to filter in ISO-8601 `YYYY-MM` format. Defaults to the current month in UTC.
          schema:
            type: string
            pattern: "^[0-9]{4}-[0-9]{2}$"
        - in: query
          name: page
          description: 1-based page number. Defaults to 1.
          schema:
            type: integer
            minimum: 1
        - in: query
          name: pageSize
          description: Page size (max 100). Defaults to 20.
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Paginated list of productivity entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListProductivitiesResponse"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createProductivity
      summary: Create a productivity entry for the authenticated user
      tags:
        - productivity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductivityRequest"
      responses:
        "201":
          description: Productivity entry created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Productivity"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/productivities/{id}:
    get:
      operationId: getProductivity
      summary: Retrieve a productivity entry by ID
      tags:
        - productivity
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Productivity entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Productivity"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProductivity
      summary: Delete a productivity entry by ID
      tags:
        - productivity
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Entry deleted successfully
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/analytics/progress:
    get:
      operationId: getProgressAnalytics
      summary: Get progress analytics for the authenticated user
      tags:
        - analytics
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: period
          description: Time period for analytics (week, month, quarter, year, all)
          schema:
            type: string
            enum: [week, month, quarter, year, all]
        - in: query
          name: category
          description: Filter by category
          schema:
            type: string
        - in: query
          name: startDate
          description: Start date in YYYY-MM-DD format
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          description: End date in YYYY-MM-DD format
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Progress analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressAnalyticsResponse"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/analytics/streak:
    get:
      operationId: getStreakAnalytics
      summary: Get streak analytics for the authenticated user
      tags:
        - analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Streak analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreakAnalyticsResponse"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/analytics/categories:
    get:
      operationId: getCategoryAnalytics
      summary: Get category breakdown analytics for the authenticated user
      tags:
        - analytics
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: startDate
          description: Start date in YYYY-MM-DD format
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          description: End date in YYYY-MM-DD format
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Category analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryAnalyticsResponse"
        "401":
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
        - code
        - message
    Productivity:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        category:
          type: string
        timeConsumedMinutes:
          type: integer
          minimum: 1
        cycleMode:
          type: string
        cycleCount:
          type: integer
          minimum: 0
        description:
          type: string
        mood:
          type: string
        imageUrl:
          type: string
          format: uri
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - category
        - timeConsumedMinutes
        - createdAt
        - updatedAt
    CreateProductivityRequest:
      type: object
      properties:
        category:
          type: string
        timeConsumedMinutes:
          type: integer
          minimum: 1
        cycleMode:
          type: string
        cycleCount:
          type: integer
          minimum: 0
        description:
          type: string
        mood:
          type: string
        imageUrl:
          type: string
          format: uri
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
      required:
        - category
        - timeConsumedMinutes
    ListProductivitiesResponse:
      type: object
      properties:
        month:
          type: string
          description: Month requested in `YYYY-MM` format.
        range:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
        data:
          type: array
          items:
            $ref: "#/components/schemas/Productivity"
        pagination:
          type: object
          properties:
            page:
              type: integer
            pageSize:
              type: integer
            totalPages:
              type: integer
            totalItems:
              type: integer
            hasNext:
              type: boolean
          required:
            - page
            - pageSize
            - totalPages
            - totalItems
            - hasNext
      required:
        - month
        - range
        - data
        - pagination
...
    ProgressAnalyticsResponse:
      type: object
      properties:
        period:
          type: string
        range:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
        stats:
          type: object
          properties:
            timeConsumedMinutes:
              type: integer
            totalSessions:
              type: integer
            totalHours:
              type: number
            mostProductiveHours:
              type: array
              items:
                type: integer
            streak:
              type: object
              properties:
                current:
                  type: integer
                longest:
                  type: integer
                lastActive:
                  type: string
                  format: date-time
            byCategory:
              type: object
              additionalProperties:
                type: integer
            byPeriod:
              type: object
              additionalProperties:
                type: object
                properties:
                  timeConsumedMinutes:
                    type: integer
                  totalSessions:
                    type: integer
                  totalHours:
                    type: number
                  byCategory:
                    type: object
                    additionalProperties:
                      type: integer
        generatedAt:
          type: string
          format: date-time
      required:
        - period
        - range
        - stats
        - generatedAt
    StreakAnalyticsResponse:
      type: object
      properties:
        current:
          type: integer
        longest:
          type: integer
        lastActive:
          type: string
          format: date-time
      required:
        - current
        - longest
        - lastActive
    CategoryAnalyticsResponse:
      type: object
      properties:
        categories:
          type: object
          additionalProperties:
            type: integer
        range:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
      required:
        - categories
        - range
...